add realationship 1-1 table to model class

+++ src/main/java/com/javacode/model/Login.java
@@ -5,6 +5,7 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import javax.persistence.OneToOne;
 import javax.persistence.Table;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Size;
@@ -15,7 +16,7 @@ public class Login {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id;
+	private Long id;
 
 	@Column(name = "user_name")
 	@NotBlank
@@ -27,7 +28,6 @@ public class Login {
 	@Size(min = 8)
 	private String password;
 
-	
 	public Login() {
 	}
 
@@ -36,7 +36,7 @@ public class Login {
 		this.password = password;
 	}
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
@@ -48,7 +48,7 @@ public class Login {
 		return password;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
