+++ src/main/java/com/javacode/repository/RegisterRepository.java
@@ -1,14 +1,12 @@
 package com.javacode.repository;
 
-import java.util.Optional;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.javacode.model.Register;
 
 @Repository
-public interface RegisterRepository extends JpaRepository<Register, Integer> {
+public interface RegisterRepository extends JpaRepository<Register, Long> {
 	
 	Register findByUserName(String userName);
